requirements.txt
aiogram==2.9.2
python-dotenv==0.14.0
pymongo==3.11.0
apscheduler==3.6.3
telegraph==1.4.1
